requirements.txt
asgiref==3.4.1
Django==3.2.8
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
Pillow==8.4.0
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
